野球必殺技改良、Com少し作成

+++ OutlowSportsFestival/source/character/Baseball/BaseballPlayerState_Counter.h
@@ -12,7 +12,7 @@
 class BaseballState_PlayerControll_Counter : public BaseballState
 {
 private:
-	int areasize;
+	float areasize;
 public:
 	BaseballState_PlayerControll_Counter(int size);
 	// ステート開始
